Removed the Spark Code from App Context as it does not help for synchronous calls.
Added spark variables to web.xml. 
1) weather static file boolean. Set true if you want the code to include
a call to the location
2) weather static file location. Set this to the path of the file HDFS
or otherwise (ensure that the HDFS does not have security blocking the
retrieval).

Added code to create a temp table called weather to the SQLContext. If
it does not persist consider making it permanent via .saveAsTable()
method.

Works within VehicleDataProcessor. Will test within aggregation function
at a later date.